+++ ExtendableEnums.EntityFrameworkCore.UnitTests/ModelBuilderExtensionsTests/ApplyExtendableEnumConversionsShould.cs
@@ -1,5 +1,4 @@
-﻿using System.Linq;
-using ExtendableEnums.EntityFrameworkCore.UnitTests.Entities;
+﻿using ExtendableEnums.EntityFrameworkCore.UnitTests.Entities;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;