+++ ExtendableEnums.LiteDB.UnitTests/BsonMapperExtensions/RegisterAllInt32BasedExtendableEnumsShould.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.IO;
-using ExtendableEnums.Testing.Models;
+﻿using ExtendableEnums.Testing.Models;
 using FluentAssertions;
 using LiteDB;
 using Microsoft.VisualStudio.TestTools.UnitTesting;