+++ ExtendableEnums.Microsoft.AspNetCore.UnitTests/ModelBindingTests.cs
@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
 using System.Net;
-using System.Net.Http;
-using System.Threading.Tasks;
 using ExtendableEnums.Testing;
 using ExtendableEnums.Testing.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;