+++ ExtendableEnums.Simple.OData.Client/ExtendableEnumConverter.cs
@@ -1,8 +1,5 @@
-﻿using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
+﻿using System.Collections.Concurrent;
 using System.Globalization;
-using System.Linq;
 using System.Reflection;
 using Simple.OData.Client;