+++ ExtendableEnums.UnitTests/ExtendableEnumTests/NewsonsoftSerializationShould.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using ExtendableEnums.Testing.Models;
+﻿using ExtendableEnums.Testing.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;