+++ ExtendableEnums.UnitTests/ExtendableEnumTests/SystemTextSerializationShould.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text.Json;
-using ExtendableEnums.Serialization.SystemText;
+﻿using System.Text.Json;
 using ExtendableEnums.Testing.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;