+++ ExtendableEnums/Serialization/Newtonsoft/ExtendableEnumJsonConverter.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using System.Globalization;
 using ExtendableEnums.Internals;
 using Newtonsoft.Json;